Make fixtures more consistent and fix spec console warnings
