+++ src/components/card/index.js
@@ -17,7 +17,7 @@ const Card = ({ body, button, image, theme, url }) => {
     <div className={classes}>
       {image && <Image src={image} />}
       {body && <div dangerouslySetInnerHTML={{ __html: body }} />}
-      {button && <Button label={button.label} url={button.url} />}
+      {button && <Button to={button.url}>{button.label}</Button>}
     </div>
   )
