+++ src/layout/footer/fixtures.js
@@ -1,10 +1,32 @@
-export const footer_test_data = {
-  policy_links: [
-    { url: "#", label: "Privacy Policy" },
-    { url: "#", label: "Terms of Service" }
-  ],
-  social_links: [
-    { url: "#", icon: "twitter" },
-    { url: "#", icon: "facebook" }
-  ]
+export default {
+  default: {
+    menus: [
+      {
+        label: "Quisque",
+        links: [
+          { url: "#", label: "Vestibulum vitae" },
+          { url: "#", label: "Pellentesque tempor venenatis" },
+          { url: "#", label: "Fusce molestie" },
+          { url: "#", label: "Etiam in lacinia dolor" }
+        ]
+      },
+      {
+        label: "Vestibulum",
+        links: [
+          { url: "#", label: "Fusce molestie" },
+          { url: "#", label: "Pellentesque venenatis" },
+          { url: "#", label: "Vestibulum vitae" },
+          { url: "#", label: "lacinia dolor" }
+        ]
+      }
+    ],
+    policy_links: [
+      { url: "#", label: "Privacy Policy" },
+      { url: "#", label: "Terms of Service" }
+    ],
+    social_links: [
+      { url: "#", icon: "twitter" },
+      { url: "#", icon: "facebook" }
+    ]
+  }
 }
