+++ src/layout/footer/link-list/fixtures.js
@@ -1,22 +1,11 @@
 export default {
-  menus: [
-    {
-      label: "Quisque",
-      links: [
-        { url: "#", label: "Vestibulum vitae" },
-        { url: "#", label: "Pellentesque tempor venenatis" },
-        { url: "#", label: "Fusce molestie" },
-        { url: "#", label: "Etiam in lacinia dolor" }
-      ]
-    },
-    {
-      label: "Vestibulum",
-      links: [
-        { url: "#", label: "Fusce molestie" },
-        { url: "#", label: "Pellentesque venenatis" },
-        { url: "#", label: "Vestibulum vitae" },
-        { url: "#", label: "lacinia dolor" }
-      ]
-    }
-  ]
+  default: {
+    label: "Quisque",
+    links: [
+      { url: "#", label: "Vestibulum vitae" },
+      { url: "#", label: "Pellentesque tempor venenatis" },
+      { url: "#", label: "Fusce molestie" },
+      { url: "#", label: "Etiam in lacinia dolor" }
+    ]
+  }
 }
