+++ src/layout/footer/social-nav/fixtures.js
@@ -1,0 +1,8 @@
+export default {
+  default: {
+    links: [
+      { url: "#", icon: "twitter" },
+      { url: "#", icon: "facebook" }
+    ]
+  }
+}
