+++ src/layout/footer/social-nav/index.js
@@ -1,33 +1,6 @@
-import React from "react"
-import PropTypes from "prop-types"
+import component from "./component"
+import fixtures from "./fixtures"
 
-import Link from "~components/link"
-import SVG from "~components/svg"
+export default component
 
-import styles from "./styles.module.scss"
-
-const SocialNav = ({ links }) => (
-  <nav className={styles.social_nav}>
-    {links.map((link, index) => (
-      <Link to={link.url} key={index}>
-        <SVG name={link.icon} />
-      </Link>
-    ))}
-  </nav>
-)
-
-SocialNav.propTypes = {
-  /**
-   * An array of social links.
-   */
-  links: PropTypes.arrayOf(
-    PropTypes.shape({
-      icon: PropTypes.string.isRequired,
-      url: PropTypes.string.isRequired
-    })
-  ).isRequired
-}
-
-SocialNav.defaultProps = {}
-
-export default SocialNav
+export { component, fixtures }
