+++ src/templates/page/test.spec.js
@@ -1,11 +1,11 @@
 import React from "react"
 import renderer from "react-test-renderer"
 
-import PageTemplate from "./"
+import { template as PageTemplate, fixtures } from "."
 
 describe("PageTemplate", () => {
   it("renders correctly", () => {
-    const tree = renderer.create(<PageTemplate sections={[]} />).toJSON()
+    const tree = renderer.create(<PageTemplate {...fixtures.default} />).toJSON()
     expect(tree).toMatchSnapshot()
   })
 })
